added dataset to delete correlated factors
I went and ran a few more models with a new datase named df3 to removed th correlated factors of eurio3m, marate, nr, and conspriceindex. the other cons variable was not highly corelated to the others so i left it in there. Camille or Dom you might want to look at the lda and other models listed towards the end to make sure the desired dataset is called. I maynot have edited like you wanted.